1/25/2024
Worked on debugging the grid system.cs . Again, I wanted to create a three dimensional grid, so unlike before to create a one level grid, you can keep the height value as zero in the inspector.
I also created the level grid as an prefab for more testing and reorganized the script for a ideally better readability.